refactor(wallet): extract hardware wallet flow and WatchOnlyTag component
