+++ apps/wallet/src/components/watch-only-tag.tsx
@@ -1,0 +1,12 @@
+import { Tag } from '@status-im/components'
+
+export function WatchOnlyTag() {
+  return (
+    <Tag
+      size="24"
+      label="Watch only"
+      aria-label="Watch-only wallet"
+      className="border-neutral-20 bg-neutral-5 text-neutral-50 hover:border-neutral-30"
+    />
+  )
+}
